+++ calpool-fe/src/pages/CarpoolManagement.css
@@ -1,6 +1,7 @@
-.CM__organization{
+.CM__organization {
     display: flex;
     flex-direction: column;
+    align-items: center;
     gap: 70px;
 }
 
@@ -8,5 +9,22 @@
     display: flex;
     flex-direction: column;
     gap: 20px;
-
+    width: 100%;
 }
+
+.edit-profile-button {
+    padding: 8px 16px;
+    margin-top: 20px; 
+    font-size: 14px; 
+    cursor: pointer;
+    border-radius: 4px;
+    border: 1px solid #ccc; 
+    background-color: transparent; 
+    transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+.edit-profile-button:hover {
+    background-color: #f5f5f5; 
+    color: #333; 
+}
+
